Refactor toss management to enhance participant tracking
