+++ examples/xmtp-group-toss/src/index.ts
@@ -1,17 +1,23 @@
 import * as fs from "fs/promises";
-import { createSigner, getEncryptionKeyFromHex } from "@helpers/client";
-import { logAgentDetails, validateEnvironment } from "@helpers/utils";
 import { HumanMessage } from "@langchain/core/messages";
 import type { createReactAgent } from "@langchain/langgraph/prebuilt";
 import {
-  Client,
+  type Client,
   type Conversation,
   type DecodedMessage,
-  type XmtpEnv,
 } from "@xmtp/node-sdk";
 import { initializeAgent, WalletService } from "./cdp";
-import { extractJsonFromResponse } from "./helper";
-import { storage, XMTP_STORAGE_DIR } from "./storage";
+import {
+  extractJsonFromResponse,
+  initializeXmtpClient,
+  TossStatus,
+  type AgentConfig,
+  type GroupTossName,
+  type MessageHandler,
+  type ParsedToss,
+  type StreamChunk,
+} from "./helper";
+import { storage } from "./storage";
 
 // Constants for default values
 const DEFAULT_OPTIONS = ["yes", "no"];
@@ -527,7 +533,7 @@ export class TossManager {
     // Validate the chosen option against available options
     if (toss.tossOptions && toss.tossOptions.length > 0) {
       const normalizedOption = chosenOption.toLowerCase();
-      const normalizedAvailableOptions = toss.tossOptions.map((opt) =>
+      const normalizedAvailableOptions = toss.tossOptions.map((opt: string) =>
         opt.toLowerCase(),
       );
 
@@ -816,107 +822,6 @@ export class TossManager {
     return toss;
   }
 }
-
-// Interface to track participant options
-export interface Participant {
-  inboxId: string;
-  option: string;
-}
-
-export interface GroupTossName {
-  id: string;
-  creator: string;
-  tossAmount: string;
-  status: TossStatus;
-  participants: string[]; // Maintaining for backward compatibility
-  participantOptions: Participant[]; // New field to track participant options
-  winner?: string;
-  walletAddress: string;
-  createdAt: number;
-  tossResult?: string;
-  paymentSuccess?: boolean;
-  transactionLink?: string;
-  tossTopic?: string;
-  tossOptions?: string[];
-}
-
-export enum TossStatus {
-  CREATED = "CREATED",
-  WAITING_FOR_PLAYER = "WAITING_FOR_PLAYER",
-  READY = "READY",
-  IN_PROGRESS = "IN_PROGRESS",
-  COMPLETED = "COMPLETED",
-  CANCELLED = "CANCELLED",
-}
-
-export interface AgentConfig {
-  configurable: {
-    thread_id: string;
-  };
-}
-
-// Interface for parsed toss information
-export interface ParsedToss {
-  topic: string;
-  options: string[];
-  amount: string;
-}
-
-// Define stream chunk types
-export interface AgentChunk {
-  agent: {
-    messages: Array<{
-      content: string;
-    }>;
-  };
-}
-
-export interface ToolsChunk {
-  tools: {
-    messages: Array<{
-      content: string;
-    }>;
-  };
-}
-
-export type StreamChunk = AgentChunk | ToolsChunk;
-
-const { WALLET_KEY, ENCRYPTION_KEY, XMTP_ENV } = validateEnvironment([
-  "WALLET_KEY",
-  "ENCRYPTION_KEY",
-  "XMTP_ENV",
-]);
-
-/**
- * Initialize the XMTP client
- */
-export async function initializeXmtpClient() {
-  // Create the signer using viem
-  const signer = createSigner(WALLET_KEY);
-  const encryptionKey = getEncryptionKeyFromHex(ENCRYPTION_KEY);
-
-  const identifier = await signer.getIdentifier();
-  const address = identifier.identifier;
-
-  const client = await Client.create(signer, encryptionKey, {
-    env: XMTP_ENV as XmtpEnv,
-    dbPath: XMTP_STORAGE_DIR + `/${XMTP_ENV}-${address}`,
-  });
-
-  logAgentDetails(address, client.inboxId, XMTP_ENV);
-
-  /* Sync the conversations from the network to update the local db */
-  console.log("✓ Syncing conversations...");
-  await client.conversations.sync();
-
-  return client;
-}
-
-export type MessageHandler = (
-  message: DecodedMessage,
-  conversation: Conversation,
-  command: string,
-) => Promise<void>;
 
 /**
  * Start listening for messages and handle them with the provided handler
